rename integration zeversolar_local zeversolar
