+++ homeassistant/components/zeversolar/const.py
@@ -1,6 +1,6 @@
 """Constants for the Local access to the zeversolar inverter integration."""
 
-DOMAIN = "zeversolar_local"
+DOMAIN = "zeversolar"
 COORDINATOR = "coordinator"
 
 ZEVER_INVERTER_ID = "zever_inverter_id"
