+++ homeassistant/components/zeversolar/sensor.py
