+++ homeassistant/components/zeversolar/strings.json
