+++ homeassistant/components/zeversolar/translations/en.json
