+++ requirements_all.txt
@@ -2455,7 +2455,7 @@ zengge==0.2
 # homeassistant.components.zeroconf
 zeroconf==0.36.1
 
-# homeassistant.components.zeversolar_local
+# homeassistant.components.zeversolar
 zeversolarlocal==0.0.9
 
 # homeassistant.components.zha
