+++ requirements_test_all.txt
@@ -1375,7 +1375,7 @@ youless-api==0.12
 # homeassistant.components.zeroconf
 zeroconf==0.36.1
 
-# homeassistant.components.zeversolar_local
+# homeassistant.components.zeversolar
 zeversolarlocal==0.0.9
 
 # homeassistant.components.zha
