+++ tests/components/zeversolar/__init__.py
